+ seperated several commands into their own cogs allowing for more modularity. implemented the "category" class attribute to compensate for the loss of organization via Cog name
